requirements.txt
Flask
Flask-Cors
tushare
pandas
gunicorn
python-dotenv